Class_03_Basic-Agent-Code_Agent-level_Run-level/pyproject.toml
[project]
name = "class-13-basic-agent-code-agent-level-run-level"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]
